Migrating the viewmodel factory
I made a new folder for viewmodel factories so that I can remove it inside my activities and ViewModel
