+++ app/src/main/java/com/example/practicecompose/viewmodel/NotesViewModel.kt
@@ -1,9 +1,14 @@
 package com.example.practicecompose.viewmodel
 
 import android.app.Application
+import android.content.Context
 import android.util.Log
 import androidx.lifecycle.ViewModel
+import androidx.lifecycle.ViewModelProvider
 import androidx.lifecycle.viewModelScope
+import androidx.room.Room
+import androidx.room.migration.Migration
+import androidx.sqlite.db.SupportSQLiteDatabase
 import com.example.practicecompose.data.room.dao.NotesDao
 import com.example.practicecompose.data.room.database.NotesDB
 import com.example.practicecompose.data.room.entity.NotesEntity
@@ -20,8 +25,12 @@ import kotlinx.coroutines.flow.update
 import kotlinx.coroutines.launch
 
 
+
+
 class NotesViewModel( private val dao : NotesDao,
 ) : ViewModel() {
+
+
 
     private val _state = MutableStateFlow(NotesState())
     private val _sortType = MutableStateFlow(Sorter.Whole)
